Refactor out custom code requirement
