+++ custom/ewsghana/models.py
@@ -1,5 +1,9 @@
 from couchdbkit.ext.django.schema import Document, BooleanProperty, StringProperty
 from casexml.apps.stock.models import DocDomainMapping
+from corehq.toggles import STOCK_AND_RECEIPT_SMS_HANDLER
+from toggle.shortcuts import update_toggle_cache, namespaced_item
+from toggle.models import Toggle
+from corehq.toggles import NAMESPACE_DOMAIN
 
 
 class EWSGhanaConfig(Document):
@@ -34,6 +38,9 @@ class EWSGhanaConfig(Document):
 
     def save(self, **params):
         super(EWSGhanaConfig, self).save(**params)
+
+        self.update_toggle()
+
         try:
             DocDomainMapping.objects.get(doc_id=self._id,
                                          domain_name=self.domain,
@@ -43,3 +50,25 @@ class EWSGhanaConfig(Document):
                                             domain_name=self.domain,
                                             doc_type='EWSGhanaConfig')
 
+    def update_toggle(self):
+        """
+        This updates the required stock and receipt sms handler
+        toggle based on whether or not EWS is enabled.
+        """
+        toggle = Toggle.get(STOCK_AND_RECEIPT_SMS_HANDLER.slug)
+        toggle_user_key = namespaced_item(self.domain, NAMESPACE_DOMAIN)
+
+        if self.enabled and toggle_user_key not in toggle.enabled_users:
+            toggle.enabled_users.append(toggle_user_key)
+            toggle.save()
+            update_toggle_cache(
+                STOCK_AND_RECEIPT_SMS_HANDLER.slug,
+                toggle_user_key, True
+            )
+        elif toggle_user_key in toggle.enabled_users:
+            toggle.enabled_users.remove(toggle_user_key)
+            toggle.save()
+            update_toggle_cache(
+                STOCK_AND_RECEIPT_SMS_HANDLER.slug,
+                toggle_user_key, False
+            )
